Allow to pass supplementary arguments and relax device constraints
Add the dpdk_supl_args argument to pass arguments such as ["--no-huge","--no-pci"] allowing to try Retina in restricted environment, i.e. with Docker using PCAP virtual device (https://doc.dpdk.org/guides/nics/pcap_ring.html).

Also, if RSS is not supported do not fail but warn the user.

Similarly, if some statistics are not available, do not fail but warn
the user.

Verify that the MTU is in the device constraint. If setting it is not
supported, do not fail but warn user.

Verify the VLAN_STRIP offload is possible. If not, just don't activate
it.