Update INSTALL.md with hugepages specifications
I am still unsure whether or not this information should be laid out in this way? The Testing in an (Offline) VM section might be a better place to put everything but since it's at the end of the document it might be less helpful? Can also switch information around and add a note in the beginning to visit that section first if choosing to work in a VM.